Make the hidden class actually hidden
